Refactored MPSArgSort
