+++ MetalSplatter/Sources/SplatRenderer.swift
@@ -3,8 +3,6 @@ import Metal
 import MetalKit
 import os
 import SplatIO
-import MetalPerformanceShaders
-import MetalPerformanceShadersGraph
 
 #if arch(x86_64)
 typealias Float16 = Float
@@ -574,7 +572,7 @@ public class SplatRenderer {
     }
 
     // Sort splatBuffer (read-only), storing the results in splatBuffer (write-only) then swap splatBuffer and splatBufferPrime
-    public func resort(useGPU: Bool = false) {
+    public func resort(useGPU: Bool = true) {
         guard !sorting else { return }
         sorting = true
         onSortStart?()
@@ -588,7 +586,7 @@ public class SplatRenderer {
 //        guard splatCount > 0 else {
 //            sorting = false
 //            let elapsed: TimeInterval = 0
-//            print("Sort time (\(useGPU ? "GPU" : "CPU")): \(elapsed) seconds")
+//            Self.log.info("Sort time (\(useGPU ? "GPU" : "CPU")): \(elapsed) seconds")
 //            onSortComplete?(elapsed)
 //            return
 //        }
@@ -662,7 +660,7 @@ public class SplatRenderer {
                 }
 
 //                let elapsed = Date().timeIntervalSince(startTime)
-//                print("Sort time (GPU): \(elapsed) seconds")
+//                Self.log.info("Sort time (GPU): \(elapsed) seconds")
 //                self.onSortComplete?(elapsed)
                 self.sorting = false
             }
@@ -706,7 +704,7 @@ public class SplatRenderer {
                 }
 
 //                let elapsedCPU = -cpuStart.timeIntervalSinceNow
-//                print("Sort time (CPU): \(elapsedCPU) seconds")
+//                Self.log.info("Sort time (CPU): \(elapsedCPU) seconds")
 //                onSortComplete?(elapsedCPU)
                 self.sorting = false
             }
@@ -795,72 +793,5 @@ private extension MTLLibrary {
             fatalError("Unable to load required shader function: \"\(name)\"")
         }
         return result
-    }
-}
-
-
-//  Original source: https://gist.github.com/kemchenj/26e1dad40e5b89de2828bad36c81302f
-//  Assessed Feb 2, 2025.
-class MPSArgSort {
-    private let dataType: MPSDataType
-    private let graph: MPSGraph
-    private let graphExecutable: MPSGraphExecutable
-    private let inputTensor: MPSGraphTensor
-    private let outputTensor: MPSGraphTensor
-
-    init(dataType: MPSDataType, descending: Bool = false) {
-        self.dataType = dataType
-
-        let graph = MPSGraph()
-        let inputTensor = graph.placeholder(shape: nil, dataType: dataType, name: nil)
-        let outputTensor = graph.argSort(inputTensor, axis: 0, descending: descending, name: nil)
-
-        self.graph = graph
-        self.inputTensor = inputTensor
-        self.outputTensor = outputTensor
-        self.graphExecutable = autoreleasepool {
-            let compilationDescriptor = MPSGraphCompilationDescriptor()
-            compilationDescriptor.waitForCompilationCompletion = true
-            compilationDescriptor.disableTypeInference()
-            return graph.compile(with: nil,
-                                 feeds: [inputTensor : MPSGraphShapedType(shape: nil, dataType: dataType)],
-                                 targetTensors: [outputTensor],
-                                 targetOperations: nil,
-                                 compilationDescriptor: compilationDescriptor)
-        }
-    }
-
-    func callAsFunction(
-        commandQueue: any MTLCommandQueue,
-        input: any MTLBuffer,
-        output: any MTLBuffer,
-        count: Int
-    ) {
-        autoreleasepool {
-            let commandBuffer = commandQueue.makeCommandBuffer()!
-            callAsFunction(commandBuffer: commandBuffer,
-                           input: input,
-                           output: output,
-                           count: count)
-            assert(commandBuffer.error == nil)
-            assert(commandBuffer.status == .completed)
-        }
-    }
-
-    private func callAsFunction(
-        commandBuffer: any MTLCommandBuffer,
-        input: any MTLBuffer,
-        output: any MTLBuffer,
-        count: Int
-    ) {
-        let shape: [NSNumber] = [count as NSNumber]
-        let inputData = MPSGraphTensorData(input, shape: shape, dataType: dataType)
-        let outputData = MPSGraphTensorData(output, shape: shape, dataType: .int32)
-        let executionDescriptor = MPSGraphExecutableExecutionDescriptor()
-        executionDescriptor.waitUntilCompleted = true
-        graphExecutable.encode(to: MPSCommandBuffer(commandBuffer: commandBuffer),
-                               inputs: [inputData],
-                               results: [outputData],
-                               executionDescriptor: executionDescriptor)
     }
 }
